feat: 점자 이미지 번역 기능 구현
