+++ src/features/brailleToHangul/index.ts
@@ -1,0 +1,1 @@
+export { translate } from './api/brailleToHangul';
